--- a/read_config_file_test.go
+++ b/read_config_file_test.go
@@ -16,7 +16,7 @@ func TestReadConfigFile(t *testing.T) {
 	var tests = []*readConfigFileTestCase{
 		{name: "Default Case", cwd: cwd, config_file: ""},
 		{name: "Specify -config", cwd: cwd, config_file: "./examples/nuxt-nuxt-js/.ls-lint.yml"},
-		{name: "Non-existent cwd", cwd: "pathThat-should_never.exist/", config_file: ".ls-lint.yml", expected_error: "open pathThat-should_never.exist/: no such file or directory"},
+		{name: "Non-existent cwd without specifying -config", cwd: "pathThat-should_never.exist/", config_file: "", expected_error: "open pathThat-should_never.exist/: no such file or directory"},
 		{name: "Non-existing config_file", cwd: cwd, config_file: "non-existent-config.yml", expected_error: "open non-existent-config.yml: no such file or directory"},
 		{name: "Non-existent config_file without specifying -config", cwd: "./npm", config_file: "", expected_error: "no config file (.ls-lint.yml or .ls-lint.yaml) was found"},
 	}
@@ -31,6 +31,10 @@ func TestReadConfigFile(t *testing.T) {
 
 		if err != nil && err.Error() != test.expected_error {
 			t.Errorf("Test %d failed (%s) with unmatched error - %s", i, test.name, err.Error())
+			return
+		}
+		if err == nil && test.expected_error != "" {
+			t.Errorf("Test %d failed (%s) - No error when an error was expected - %s", i, test.name, test.expected_error)
 			return
 		}