--- a/test.go
+++ b/test.go
@@ -5,3 +5,10 @@ type ruleTest struct {
 	expected bool
 	err      error
 }
+
+type readConfigFileTestCase struct {
+	name					string
+	cwd						string
+	config_file 	string
+	expected_error string
+}