Should handle #29 if you want to use `.ls-lint.yaml` instead. Also allows for specifying a config file `ls-lint -config ./relative/path/to/config/.ls-lint.yaml`.

You can specify the relative path to a new working directory with `-pwd` flag. This could be useful for multi-language monorepos where projects have different directory naming conventions. `ls-lint -pwd ./relative/path -config ./relative/path/.ls-lint.yml`

Hyperfine performance results (2022/10/07 1:30 PM MT)

| build | results |
| ----- | ------ |
| master | `Time (mean ± σ):      85.0 ms ±   1.8 ms    [User: 48.4 ms, System: 32.4 ms]` | 
| this branch | `Time (mean ± σ):      84.6 ms ±   2.1 ms    [User: 48.4 ms, System: 32.2 ms]` |

Run on:
```
Hardware Overview:

      Model Name: MacBook Pro
      Model Identifier: MacBookPro16,1
      Processor Name: 6-Core Intel Core i7
      Processor Speed: 2.6 GHz
      Number of Processors: 1
      Total Number of Cores: 6
      L2 Cache: 768 KB
      L3 Cache: 6 MB
      Hyper-Threading Technology: Enabled
      Memory: 16 GB
```

I didn't see any contribution guidelines, so let me know if you'd like me to do something else.