@loeffel-io 2022-10-07
thanks @Mctalian for the PR! To specify a config file should not change the root - maybe there should be another option for this? 🤔