@Mctalian 2022-10-07
> 



> thanks @Mctalian for the PR! To specify a config file should not change the root - maybe there should be another option for this? 🤔

Sure! I can look into changing that.

Any tips on how to get past the codecov checks? I can't test the `main` func so I'm a bit stuck 😅