@Mctalian 2022-10-07
@loeffel-io I added the `-pwd` flag which will change the root. `-config` no longer changes the root.

I moved the logging of errors out of main to try to increase coverage - it did increase, but not enough to pass the patch check. Not sure how much more I can increase that coverage 🙁