@loeffel-io 2023-05-21
With `v2.x` this is provided by the `config` and `workdir` option: https://github.com/loeffel-io/ls-lint/releases/tag/v2.0.0-beta.0

Feel free to try it out